Started moving code around to separate source and tests.
