+++ include/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(gtest_ct INTERFACE)
+target_include_directories(gtest_ct INTERFACE .)
+
